requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
psycopg==3.2.9
psycopg-binary==3.2.9
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.15.0
whitenoise==6.9.0
